requirements.txt
# Google Drive
google-api-python-client==2.108.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# PDF Processing
pdf2image==1.17.0
Pillow==10.2.0
pytesseract==0.3.10
pypdf==4.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3

# OpenAI
openai==1.3.0

# Dashboard
streamlit==1.28.0
streamlit-authenticator==0.2.3
pandas==2.1.3
plotly==5.18.0
openpyxl==3.1.2
PyYAML==6.0.1

# Logging
python-json-logger==2.0.7

# Security
bcrypt==4.1.1

